Generate Composite Scores

This function calculates composite scores based on subtest correlations and their inter-correlations.

Usage

composite(
  r,
  r.between,
  mean = 10,
  sd = 3,
  comp.mean = 100,
  comp.sd = 15,
  dp = 2
)

Arguments

r

A numeric vector of correlations for the included tests.

r.between

A numeric vector of inter-correlations between the included tests.

mean

A numeric value specifying the mean of the included tests. Default is 10.

sd

A numeric value specifying the standard deviation of the included tests. Default is 3.

comp.mean

A numeric value specifying the mean of the composite score. Default is 100.

comp.sd

A numeric value specifying the standard deviation of the composite score. Default is 15.

dp

An integer specifying the number of decimal places for rounding results. Default is 2

Value

A list of class 'gen_comp' containing:

k

Number of subtests.

r

Vector of correlations for the included tests.

r.between

Vector of inter-correlations between the included tests.

composite_r

Composite correlation.

original_mean

Original mean of the composite score.

original_sd

Original standard deviation of the composite score.

original_sem

Original standard error of measurement.

original_semt

Original standard error of the mean transformed.

transformed.mean

Transformed mean of the composite score.

transformed.sd

Transformed standard deviation of the composite score.

transformed.sem

Transformed standard error of measurement.

transformed.semt

Transformed standard error of the mean transformed.

Examples

result <- composite(r = c(.87, .94, .94),
                             r.between = c(0.74, 0.64, 0.73),
                             dp = 3)
print(result)
#> Generate Composite Score
#> 
#> INPUTS:
#> 
#> 
#> Inputs                     Value            
#> -------------------------  -----------------
#> Number of tests included   3                
#> r for included tests       0.87, 0.94, 0.94 
#> r between included tests   0.74, 0.64, 0.73 
#> 
#> OUTPUTS:
#> 
#> 
#> Outputs       Original   Transformed 
#> ------------  ---------  ------------
#> Composite r   0.9650                 
#> Mean          30.0000    100.0000    
#> SD            7.4820     15.0000     
#> SEM           1.3920     2.7910      
#> SEMt (true)   1.3440     2.6950