Generate Regression Norms for a Single Variable

This function generates regression norms and performs significance testing for an individual's score on a predictor variable. It calculates the regression equation, standard errors, and significance levels based on summary statistics from a control sample.

Usage

cont_norms_single_pred(
  ctrl.x.mean,
  ctrl.x.sd,
  ctrl.y.mean,
  ctrl.y.sd,
  r,
  n,
  x,
  y,
  conf.level = 0.05,
  direction = "lower",
  dp = 4,
  x.name = "X",
  y.name = "Y"
)

Arguments

ctrl.x.mean

Mean of the predictor variable in the control sample.

ctrl.x.sd

Standard deviation of the predictor variable in the control sample.

ctrl.y.mean

Mean of the criterion variable in the control sample.

ctrl.y.sd

Standard deviation of the criterion variable in the control sample.

r

Correlation coefficient between the predictor and criterion variables in the control sample.

n

Sample size of the control group.

x

Score of the individual on the predictor variable.

y

Obtained score of the individual on the criterion variable.

conf.level

Confidence level for the significance test. Default is 0.05.

direction

Direction for the significance test. Options are "lower" or "upper". Default is "lower".

dp

Number of decimal places.

x.name

Name for the predictor variable. Default is "X".

y.name

Name for the criterion variable. Default is "Y".

Value

A list containing calculated values and results.

Examples

cont_norms_single_pred(
  ctrl.x.mean = 63.8,
  ctrl.x.sd = 8.42,
  ctrl.y.mean = 41.3,
  ctrl.y.sd = 13.2,
  r = -0.58,
  n = 160,
  x = 26,
  y = 41,
  x.name = "Predictor",
  y.name = "Criterion"
)
#> Neuropsychological Regression Norms Single Case Analysis
#> 
#> Regression norms and significance testing for an individual case based on the control sample.
#> 
#> INPUTS:
#> 
#> Variable                 Mean      SD   Case's Score  n     r        t1 Score
#> ----------------------  -----  ------  -------------  ----  ------  ---------
#> Predictor (Predictor)    63.8    8.42             26  160   -0.58          26
#> Criterion (Criterion)    41.3   13.20             41                       41
#> 
#> OUTPUTS:
#> 
#> Outputs                                     Value                        
#> ------------------------------------------  -----------------------------
#> Regression equation (Predicted Y)           99.311 + -0.9093 * Predictor 
#> Standard error of estimate                  10.7869                      
#> Case's PREDICTED score                      75.6702                      
#> Discrepancy (Observed - Predicted)          -34.6702                     
#> Effect size (Z-OP)                          -3.2243                      
#> t value                                     -3.0196                      
#> One-tailed p-value                          0.0015                       
#> Two-tailed p-value                          0.0030                       
#> Estimated % with more extreme discrepancy   0.1476 %                     
#> 
#> Based on: Crawford, Garthwaite, & Porter (2010), Crawford & Garthwaite (2002), and Crawford & Howell (1998).