Assessing for a Frequentist Discrepancy Between Two Test Scores for an Individual Case.

This function offers a range of approaches for assess whether the discrepancy in scores between two tests for a patient is significantly larger than that which would be expected for a control or normative sample. The analysis yields a significance test, a point estimate of the abnormality of the score difference, and confidence intervals for the abnormality of the difference.

Usage

discrep(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  direction.x = "lower",
  direction.y = "lower",
  tail = "one.tailed",
  test.names = c("test X", "test Y"),
  conf.level = 0.95,
  dp = 2,
  method = "rsdt"
)

Arguments

ctrl.mean.x

Mean of the control group for the first test.

ctrl.sd.x

Standard deviation of the control group for the first test.

ctrl.mean.y

Mean of the control group for the second test.

ctrl.sd.y

Standard deviation of the control group for the second test.

ctrl.r.xy

Correlation between the two tests in the control group.

ctrl.n

Size of the control group.

score.x

Test score of the individual case for the first test.

score.y

Test score of the individual case for the second test.

direction.x

Specifies the expected direction for the first test score. Use "lower" if a lower score indicates worse performance or "higher" if a higher score indicates worse performance (default: "lower").

direction.y

Specifies the expected direction for the second test score. Use "lower" if a lower score indicates worse performance or "higher" if a higher score indicates worse performance (default: "lower").

tail

Character. Specifies whether the test is one-tailed or two-tailed. Options are "one.tailed" and "two.tailed" (default)

test.names

A vector of two strings representing the names of the tests (default is c("X", "Y")).

conf.level

Confidence level (default is 0.95 for 95%).

dp

Number of decimal places for rounding the results (default is 2).

method

Character string specifying the method used to calculate the discrepancy: "rsdt" (revised standardized difference test), "usdt" (uncorrected standardized difference test), or "difflims" (difference limits method). Default is "rsdt".

Value

A list containing statistical inputs, parameters, and results. Key outputs include:

t value: The t-value calculated for the test.

p value: The p-value for the test, indicating statistical significance.

effect size (z-cc): The z-score (effect size) corrected for the control group.

abnormality: The percentage of the population expected to score a more extreme score.

Details

This function assesses a dissociation between two test scores for a single case using a range of methods.

The modified paired samples t-test approach of Crawford et al. (1998) is used by comparison to a normative sample. Unlike earlier methods (e.g., Payne & Jones), this method treats data from the normative sample as sample statistics and not population parameters. The result provided is a t score and associated p value. This approach helps to address the problem associated with small normative samples.

RSDT

USDT

check this... "rsdt" (revised standardized difference test), "usdt" (uncorrected standardized difference test), or "difflims" (difference limits method).

In addition to determining whether a difference exists, it is also important to understand the magnitude of that difference. Therefore, it is often recommended that effect sizes are provided alongside p-values to estimate the size of the observed effect. To this effect, Crawford et al. (1998) provided a method for deriving an effect size in single-case studies using the case-controls design (z-cc), where a single patient's cognitive performance is compared to a matched control group. The modified z-score (z-cc) is provided as both point and interval estimates.

Finally, neuropsychologists often need to determine how abnormal a patient's test score is. In the case of the modified t-test, the abnormality can be easily estimated by multiplying the t-value by 100 (Crawford & Howell, 1998). This estimate quantifies the percentage of the population expected to exhibit a more extreme score. Confidence limits on the estimate of abnormality are also provided (Crawford & Garthwaite, 2002).

References

Crawford, J.R., & Garthwaite, P.H. (2002). Investigation of the single case in neuropsychology: confidence limits on the abnormality of test scores and test score differences. Neuropsychologia, 40(2002), 1196–1208.

Crawford, J.R., Howell, D.C., & Garthwaite, P.H. (1998). Payne and Jones Revisited: Estimating the Abnormality of Test Score Differences Using a Modified Paired Samples t Test. Journal of Clinical and Experimental Neuropsychology, 20(6), 898-905.

Crawford, J.R., & Howell, D.C. (1998). Comparing an individual’s test score against norms derived from small samples. The Clinical Neuropsychologist, 12(4), 482-486.

Crawford, J.R., Garthwaite, P.H., & Porter, S. (2010). Point and interval estimates of effect sizes for the case-controls design in neuropsychology: Rationale, methods, implementations, and proposed reporting standards. Cognitive Neuropsychology, 27(3), 245-260.

Payne, R. W., & Jones, G. (1957). Statistics for the investigation of individual cases. Journal of Clinical Psychology, 13, 115-121.

See also

Examples

discrep(100, 15, 110, 10, 0.5, 30, 130, 120, direction.x = "lower",
  direction.y = "higher", test.names = c("Fluency", "Sequencing"))
#> Testing for a Frequentist Discrepency Between Two Test Scores Compared to a Control Sample.
#> 
#> INPUTS:
#> 
#> Test          Mean   SD  Sample size   r      Case score
#> -----------  -----  ---  ------------  ----  -----------
#> Fluency        100   15  30            0.5           130
#> Sequencing     110   10                              120
#> 
#> PARAMETERS:
#> 
#> Parameter                                       Value                                    
#> ----------------------------------------------  -----------------------------------------
#> Discrepancy Method                              RSDT (Crawford & Garthwaite, 2005)       
#> Confidence Interval Method                      Modified T (Crawford & Garthwaite, 2002) 
#> Confidence Intervals                            95%                                      
#> Hypothesis                                      One Tailed                               
#> Direction Indicating Impairment (Fluency):      Lower                                    
#> Direction Indicating Impairment (Sequencing):   Higher                                   
#> 
#> OUTPUTS:
#> 
#> Variable                                             Value    95% CI           
#> ---------------------------------------------------  -------  -----------------
#> Effect size (z) for  Fluency                         2.00                      
#> Effect size (z) for  Sequencing                      -1.00                     
#> Effect size (z-dcc) between Fluency and Sequencing   3.00     2.15 to 3.84     
#> t value                                              2.87                      
#> p value                                              0.00                      
#> Abnormality                                          0.38 %   0.01 % to 1.59 % 
#> 
#> Note.
#> - Abnormality = The percentage of controls expected to show a higher deficit.
#> - z-cc = Z for the case control.
#> 
#> See documentation for further information on how scores are computed and how to cite methods.