USDT

description.

Usage

discrep_usdt(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  direction = "lower",
  tail = "one.tailed",
  test.names = c("test X", "test Y"),
  conf.level = 0.95,
  dp = 2
)

Arguments

ctrl.mean.x

Mean of the control group for the first test.

ctrl.sd.x

Standard deviation of the control group for the first test.

ctrl.mean.y

Mean of the control group for the second test.

ctrl.sd.y

Standard deviation of the control group for the second test.

ctrl.r.xy

Correlation between the two tests in the control group.

ctrl.n

Integer value representing the sample size of the control group.

score.x

Test score of the individual case for the first test.

score.y

Test score of the individual case for the second test.

direction

Character. Specifies the direction of the hypothesis. Options are "lower" (one-tailed), "higher" (one-tailed), or "two.tailed" (default, two-tailed).

tail

Character. Specifies whether the test is one-tailed or two-tailed. Options are "one.tailed" and "two.tailed" (default)

test.names

A vector of two strings representing the names of the tests (default is c("X", "Y")).

conf.level

Confidence level (default is 0.95 for 95%).

dp

Number of decimal places for rounding the results (default is 2).

Value

A list of statistical input, parameters, and results. Key outputs include:

t value: The t-value calculated for the test.

p value: The p-value for the test, indicating statistical significance.

effect-size (z-cc): The z-score (effect-size) corrected for the control group.

abnormality: The percentage of the population expected to score a more extreme score.

Details

...

References

To complete

To complete

See also

Examples

discrep_usdt(100, 15, 110, 10, 0.5, 30, 130, 120, test.names = c("Fluency", "Sequencing"))
#> Frequentist Dissociation Between Two Test Scores Compared to a Control Sample (USDT Approach).
#> 
#> INPUTS:
#> 
#> Test          Mean   SD  Sample size   r      Case score
#> -----------  -----  ---  ------------  ----  -----------
#> Fluency        100   15  30            0.5           130
#> Sequencing     110   10                              120
#> 
#> OUTPUTS:
#> 
#> Variable                                             Value     95% Confidence Interval 
#> ---------------------------------------------------  --------  ------------------------
#> Effect size (z) for  Fluency                         2.00                              
#> Effect size (z) for  Sequencing                      1.00                              
#> Effect size (z-dcc) between Fluency and Sequencing   1.51      0.98 to 2.03            
#> t value                                              1.49                              
#> p-value                                              0.93                              
#> Abnormality                                          92.65 %   83.61 % to 97.9 %       
#> 
#> See documentation for further information on how scores are computed.