Dissociation Analysis Between Two Test Scores

This function assesses dissociations between two test scores (X and Y) compared to control data. It helps determine if a single case's scores on two tests are abnormally divergent relative to a control sample. Both frequentist (t-test) and Bayesian approaches are available for analysis.

Usage

dissociation(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  direction.x = "lower",
  direction.y = "lower",
  tail = "one.tailed",
  test.names = c("test X", "test Y"),
  conf.level = 0.95,
  dp = 2,
  deficit_method = "t",
  discrep_method = "rsdt"
)

Arguments

ctrl.mean.x

Mean of the control group for the first test.

ctrl.sd.x

Standard deviation of the control group for the first test.

ctrl.mean.y

Mean of the control group for the second test.

ctrl.sd.y

Standard deviation of the control group for the second test.

ctrl.r.xy

Correlation between the two tests in the control group.

ctrl.n

Integer value representing the sample size of the control group.

score.x

Test score of the individual case for the first test.

score.y

Test score of the individual case for the second test.

direction.x

Specifies the expected direction for the first test score. Use "lower" if a lower score indicates worse performance or "higher" if a higher score indicates worse performance (default: "lower").

direction.y

Specifies the expected direction for the second test score. Use "lower" if a lower score indicates worse performance or "higher" if a higher score indicates worse performance (default: "lower").

tail

Character. Specifies whether the test is one-tailed or two-tailed. Options are "one.tailed" and "two.tailed" (default)

test.names

A vector of two strings representing the names of the tests (default is c("X", "Y")).

conf.level

Confidence level (default is 0.95 for 95%).

dp

Number of decimal places for rounding the results (default is 2).

deficit_method

Character. Method for the deficit analysis. Options are 't' for frequentist t-test and 'bayes' for Bayesian analysis.

discrep_method

Character. Method for the discrepancy analysis. Default is 'rsdt' (revised standardized difference test).

Value

A list containing the results of the dissociation analysis. Key outputs include:

x.res and y.res: Results for each test score, including t-values, p-values, effect sizes, and abnormality percentages.

discrepancy.res: Results of a discrepancy analysis between the two test scores (if applicable).

References

Crawford, J.R., & Garthwaite, P.H. (2007). Comparison of a single case to a control or normative sample in neuropsychology: Development of a Bayesian approach. Cognitive Neuropsychology, 24(4), 343-372.

See also

Examples

# Example usage
dissociation(
  ctrl.mean.x = 100, ctrl.sd.x = 15, ctrl.mean.y = 105, ctrl.sd.y = 10,
  ctrl.r.xy = 0.8, ctrl.n = 30, score.x = 130, score.y = 90
)
#> Testing for a Frequentist Dissociation Between Two Test Scores Compared to a Control Sample.
#> 
#> INPUTS:
#> 
#> Test      Mean   SD  Sample size   r      Case score
#> -------  -----  ---  ------------  ----  -----------
#> test X     100   15  30            0.8           130
#> test Y     105   10                               90
#> 
#> PARAMETERS:
#> 
#> item                                       value                                    
#> -----------------------------------------  -----------------------------------------
#> Deficit Method                             Modified T (Crawford & Howell, 1998)     
#> Discrepancy Method                         RSDT (Crawford & Garthwaite, 2005)       
#> Confidence Interval Method                 Modified T (Crawford & Garthwaite, 2002) 
#> Confidence Intervals                       95%                                      
#> Hypothesis                                 One-Tailed                               
#> Direction Indicating Impairment (test X)   Lower                                    
#> Direction Indicating Impairment (test Y)   Lower                                    
#> 
#> OUTPUTS:
#> 
#> 1) DEFICIT ANALYSIS:
#> 
#> Test      t-value   p-value   z-dcc  95% CI            Abnormality  95% CI           Deficit 
#> -------  --------  --------  ------  ---------------  ------------  ---------------  --------
#> test X       1.97      0.97     2.0  1.37 to 2.62            97.06  91.45 to 99.56   FALSE   
#> test Y      -1.48      0.08    -1.5  -2.02 to -0.97           7.54  2.18 to 16.63    FALSE   
#> 
#> 2) DISCREPANCY ANALYSIS:
#> 
#> Statistic                                       Value    95% CI       
#> ----------------------------------------------  -------  -------------
#> Effect size (z) for  test X                     2.00                  
#> Effect size (z) for  test Y                     -1.50                 
#> Effect size (z-dcc) between test X and test Y   5.53     4.07 to 6.95 
#> t-value                                         5.21                  
#> p-value                                         0.00                  
#> Abnormality                                     0.00 %   0 % to 0 %   
#> 
#> 3) DISSOCIATION ANALYSIS:
#> 
#> No significant dissociation was found between the test scores.
#> 
#> Note.
#> - Abnormality = The percentage of controls expected to show a higher deficit.
#> - z-dcc = Z  discrepancy for the case control.
#> 
#> See documentation for further information on how scores are computed.