Calculate abnormality of two test scores with confidence intervals using the modified t-test.

Calculate abnormality of two test scores with confidence intervals using the modified t-test.

Usage

dissociation_double(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  conf.level = 0.95,
  direction = "lower",
  dp = 4,
  test.names = c("X", "Y")
)

Arguments

ctrl.mean.x

Mean of the control group for the first test.

ctrl.sd.x

Standard deviation of the control group for the first test.

ctrl.mean.y

Mean of the control group for the second test.

ctrl.sd.y

Standard deviation of the control group for the second test.

ctrl.r.xy

Correlation between the two tests in the control group.

ctrl.n

Size of the control group.

score.x

Test score of the individual for the first test.

score.y

Test score of the individual for the second test.

conf.level

Confidence level for the interval (default is 0.05).

direction

Direction of the test, either "lower" or "higher" (default is "lower").

dp

Number of decimal places for rounding the results (default is 2).

test.names

A vector of two strings representing the names of the tests (default is c("X", "Y")).

Value

A list containing the z-scores for both tests, control group means and standard deviations, correlation between tests, control group size, scores for both tests, confidence level, direction of the test, t-value, p-value, two-tailed p-value, zdcc, zdcc confidence interval lower bound, zdcc confidence interval upper bound, abnormality percentage, abnormality confidence interval lower bound, and abnormality confidence interval upper bound.

Examples

dissociation_double(100, 15, 110, 10, 0.5, 30, 130, 120, test.names = c("Test A", "Test B"))
#> Frequentist Double Dissociation
#> 
#> Frequentist point estimate and confidence limits on the abnormality of test score differences within a single case.
#> 
#> INPUTS:
#> 
#> Test      Mean   SD  Sample size   r      Case score
#> -------  -----  ---  ------------  ----  -----------
#> Test A     100   15  30            0.5           130
#> Test B     110   10                              120
#> 
#> OUTPUTS:
#> 
#> Variable                                        Value       95% Confidence Interval 
#> ----------------------------------------------  ----------  ------------------------
#> Effect size (z) for test Test A                 2.000                               
#> Effect size (z) for test Test B                 1.000                               
#> Effect size (z-dcc) between Test A and Test B   1.000       0.5541 to 1.4343        
#> t value                                         0.984                               
#> One-tailed p-value                              0.8334                              
#> Two-tailed p-value                              0.3333                              
#> Abnormality                                     83.3373 %   71.0259 % to 92.4261 %  
#> 
#> Based on: Crawford, Garthwaite & Howell (1998).