Calculate the percentage of the normative population with j or more abnormal test scores

This function calculates the percentage of the normative population that would exhibit j or more abnormal test scores in a battery of tests. The method uses multivariate normal simulation.

Usage

prevalence_of_j_abnormal_scores(R, sims = 1e+05, threshold = -1.645)

Arguments

R

A correlation matrix for the tests.

sims

Number of simulations to perform. Default is 100000.

threshold

Threshold for abnormal test scores. Default is -1.645 (corresponding to the 5th percentile in a standard normal distribution).

Value

A list containing:

props

Proportion of individuals with more than j-1 abnormal scores.

props.two.tail

Proportion of individuals with more than j-1 abnormal scores, adjusted for a two-tailed test.

Examples

# Example usage with a 4x4 identity matrix as the correlation matrix
R <- diag(4)
prevalence_of_j_abnormal_scores(R)
#> Abnormality of j abnormal scores in a battery of tests.
#> 
#> Parameters:
#> 
#> k:                         4
#> Number of simulations:     1e+05
#> Abnormality threshold: z = -1.645
#> 
#> Output (One-Tailed):
#> 
#> Proportion of people with more than:
#> 0 abnormal scores: 18.56%
#> 1 abnormal scores: 1.38%
#> 2 abnormal scores: 0.05%
#> 3 abnormal scores: 0.00%
#> 
#> Output (Two-Tailed):
#> 
#> Proportion of people with more than:
#> 0 abnormal scores: 37.13%
#> 1 abnormal scores: 2.75%
#> 2 abnormal scores: 0.10%
#> 3 abnormal scores: 0.00%