Compare Stability Methods

This function compares stability methods given the scores of two tests.

Arguments

t1.score

Scores of the first test.

t2.score

Scores of the second test.

norm.t1.mean

Mean of the normative sample at time 1.

norm.t1.sd

Standard deviation of the normative sample at time 1.

norm.t2.mean

Mean of the normative sample at time 2.

norm.t2.sd

Standard deviation of the normative sample at time 2.

norm.r

Reliability of the normative sample.

norm.n

Sample size of the normative sample.

conf.level

Confidence level for stability assessment.

methods

A character vector specifying which methods to calculate. Default is "all" methods.

Examples

# Example data
t1.score <- 88
norm.t1.mean <- 113
norm.t1.sd <- 7.2
norm.t2.mean <- 118.7
norm.t2.sd <- 8.3
norm.r <- 0.93
norm.n <- 821
t2.score <- 103
ci <- 0.95

# Using all methods
result_all <- reliable_change_discreps(t1.score, t2.score,
norm.t1.mean, norm.t1.sd, norm.t2.mean, norm.t2.sd, norm.r, norm.n,
ci)
print(result_all)
#> Comparison of Stability Methods
#> 
#> INPUT:
#> 
#> Variable     Mean    SD   Case's Score  n     r    
#> ---------  ------  ----  -------------  ----  -----
#> Time 1      113.0   7.2             88  821   0.93 
#> Time 2      118.7   8.3            103             
#> 
#> 
#> PARAMS:
#> Confidence level: 95%
#> 
#> OUTPUT:
#> 
#> Method      Predicted   Observed   Discrepency   Error   CI lb    CI ub      Z
#> ---------  ----------  ---------  ------------  ------  ------  -------  -----
#> jacobson        88.00        103         15.00    2.69   82.73    93.27   5.58
#> speer           95.45        103          7.55    2.69   90.18   100.72   2.81
#> chelune         93.70        103          9.30    2.69   88.43    98.97   3.46
#> mcsweeny        91.90        103         11.10    3.05   85.92    97.88   3.64
#> charter         95.45        103          7.55    3.05   89.47   101.43   2.48
#> CH              91.90        103         11.10    3.07   85.88    97.92   3.62
#> temkin          93.70        103          9.30    3.09   87.64    99.76   3.01
#> iverson         93.70        103          9.30    2.91   88.00    99.40   3.20
#> maassen         89.88        103         13.12    2.91   84.18    95.58   4.51
#> crawford        89.88        103         13.12    2.93   84.14    95.62   4.48

# Using selected methods
result_selected <- reliable_change_discreps(t1.score, t2.score,
norm.t1.mean, norm.t1.sd,norm.t2.mean, norm.t2.sd, norm.r,
norm.n, ci, methods = c("crawford", "maassen"))
print(result_selected)
#> Comparison of Stability Methods
#> 
#> INPUT:
#> 
#> Variable     Mean    SD   Case's Score  n     r    
#> ---------  ------  ----  -------------  ----  -----
#> Time 1      113.0   7.2             88  821   0.93 
#> Time 2      118.7   8.3            103             
#> 
#> 
#> PARAMS:
#> Confidence level: 95%
#> 
#> OUTPUT:
#> 
#> Method      Predicted   Observed   Discrepency   Error   CI lb   CI ub      Z
#> ---------  ----------  ---------  ------------  ------  ------  ------  -----
#> crawford        89.88        103         13.12    2.93   84.14   95.62   4.48
#> maassen         89.88        103         13.12    2.91   84.18   95.58   4.51