Standard Error of Difference Between Means

Calculate the standard error of the difference (SED) between means.

Usage

se_difference(sd.1, sd.2, r)

Arguments

sd.1

(numeric) Standard deviation of the first group or vector of standard deviations for each observation in the first group.

sd.2

(numeric) Standard deviation of the second group or vector of standard deviations for each observation in the second group.

r

(numeric) The reliability of the test used to obtain the standard deviations.

Value

The standard error of difference between means.

Details

The SED is the expected spread of the distribution of change scores if no actual change had occurred. This is calculated using the approach offered by Iverson (2001) which uses the SEM of the normative data for each testing interval.

Examples

# If sd.1 and sd.2 are single values
se_difference(sd.1 = 3, sd.2 = 4, r = 0.5)
#> [1] 3.535534

# If sd.1 and sd.2 are vectors of the same length
se_difference(sd.1 = c(3, 4, 5), sd.2 = c(4, 5, 6), r = 0.5)
#> [1] 1