Standard Error of Measurement (SEM)

Calculate the standard error of measurement (SEM) given the reliability and standard deviation of scores

Usage

se_measurement(dat = NULL, sd = NULL, r = NULL)

Arguments

dat

(numeric) A numeric vector of measurements from which to calculate the standard deviation. Ignored if r and SD are provided.

sd

(numeric) The standard deviation of the measurements. If dat is provided, sd will be calculated from dat.

r

(numeric) The reliability of the test.

Value

The standard error of measurement (SEM).

Details

The Standard Error of Measurement is the standard deviation of an infinite number of hypothetical scores around the true score. The SEM is calculated as the standard deviation (SD) multiplied by the square root of 1 minus the test retest reliability (r) of the measurement tool.

Examples

# If dat and r are provided, calculate sd from dat and then SEM
se_measurement(dat = c(1, 2, 3, 4, 5), r = 0.4)
#> [1] 1.224745

# If sd and r are provided, calculate SEM
se_measurement(sd = 3, r = 0.4)
#> [1] 2.32379