Plot Stability

Plot Stability

Usage

plot_change(
  data,
  metric = "z",
  group,
  test,
  t1.score,
  t2.expected,
  t2.score,
  t2.ci.lb,
  t2.ci.ub,
  axis.label.metric = metric,
  abbreviations = TRUE
)

Arguments

data

The data frame containing the input data.

metric

The metric to be used ("z" by default).

group

The grouping variable in the data.

test

The test variable in the data.

t1.score

The score at time point 1.

t2.expected

The expected score at time point 2.

t2.score

The actual score at time point 2.

t2.ci.lb

The lower bound of the confidence interval for time point 2.

t2.ci.ub

The upper bound of the confidence interval for time point 2.

axis.label.metric

The metric used for axis labeling (default is "z").

abbreviations

Logical, whether to abbreviate test names (default is TRUE).

Value

A ggplot object representing the stability plot.