Estimate Neuropsychological Test Scores Based on Earlier Administrations

This function calculates predicted values for neuropsychological tests based on scores from an earlier administration along with stability data (M, SD, test-retest) from a normative sample. It allows the user to determine if the change in a score between two time points is statistically significant.

Usage

predict_t2(
  method = "crawford",
  t1.score,
  norm.t1.mean,
  norm.t1.sd,
  norm.t2.mean,
  norm.t2.sd,
  norm.r,
  norm.n,
  conf.level = 0.95,
  dp = 2
)

Arguments

method

Character string specifying the method used for reliable change estimate calculation. Options include "crawford", "jacobson", "speer", "chelune", "mcsweeny", "charter", "CH", "temkin", "iverson", and "maassen". Default is "crawford".

t1.score

Numeric value of the score at time 1.

norm.t1.mean

Numeric value of the mean from the normative data at time 1.

norm.t1.sd

Numeric value of the standard deviation from the normative data at time 1.

norm.t2.mean

Numeric value of the mean from the normative data at time 2.

norm.t2.sd

Numeric value of the standard deviation from the normative data at time 2.

norm.r

Numeric value of the Pearson correlation coefficient between time 1 and time 2 scores.

norm.n

Numeric value of the sample size of the normative data.

conf.level

Confidence level for stability assessment.

dp

Integer specifying the number of decimal places to round the converted scores to. Default is 2.

Value

A list containing the following components:

method: The method used for calculation.

t1.score: The input score at time 1.

t2.expected: The expected score at time 2 based on the method used.

t2.sem: The standard error of measurement for the time 2 score.

t2.ci.lb: The lower bound of the confidence interval for the expected score at time 2.

t2.ci.ub: The upper bound of the confidence interval for the expected score at time 2.

Details

Note that this function is an intermediary function and that the function XXXX will be more suited to estimating and analysing stability or change.

The function supports multiple methods for calculating reliable change, including those proposed by Crawford & Garthwaite (2006), Jacobson & Truax (1991), Speer (1992), Chelune et al. (1993), Iverson et al. (2001), McSweeny et al. (1993), Maassen et al. (2006), and Charter (1996). By default, it uses the Crawford method (Crawford & Garthwaite, 2006).

The available methods differ in their approach to estimating the predicted value and the associated standard error. While all of the included approaches account for reliability of the test (i.e,. test-retest), only some of the approaches account for practice effects and regression to the mean. Readers who want to consider the merits of each approach are invited to read the introductory papers provided by Hinton-Bayre (2010) and Duff (2012).

References

Jacobson, N.S., & Truax, P. (1991). Clinical significance: A statistical approach to defining meaningful change in psychotherapy research. Journal of Consulting and Clinical Psychology, 59(1), 12-19.

Duff, K. (2012). Evidence-Based Indicators of Neuropsychological Change in the Individual Patient: Relevant Concepts and Methods. Archives of Clinical Neuropsychology, 27(3), 248–261. doi:10.1093/arclin/acr120. PMCID: PMC3499091.

Hinton-Bayre, A. D. (2010). Deriving Reliable Change Statistics from Test–Retest Normative Data: Comparison of Models and Mathematical Expressions. Archives of Clinical Neuropsychology, 25(3), 244–256. doi:10.1093/arclin/acq008.

Crawford, J. R., & Garthwaite, P. H. (2006). Comparing patients’ predicted test scores from a regression equation with their obtained scores: A significance test and point estimate of abnormality with accompanying confidence limits. Neuropsychology, 20, 259-271.

Examples

# Calculate reliable change using the Jacobson method
predict_t2(
  method = "jacobson",
  t1.score = 100,
  norm.t1.mean = 103,
  norm.t1.sd = 10,
  norm.t2.mean = 109,
  norm.t2.sd = 8,
  norm.r = 0.76,
  norm.n = 100,
  conf.level = 0.95
)
#> Estimate Neuropsychological Test Scores Based on Earlier Administrations.
#> 
#> INPUT:
#> 
#> Variable    Mean   SD  Case's Score   n     r    
#> ---------  -----  ---  -------------  ----  -----
#> Time 1       103   10  100            100   0.76 
#> Time 2       109    8                            
#> 
#> PARAMS:
#> 
#> Change method:    jacobson
#> Confidence level: 0.95
#> 
#> OUTPUT:
#> 
#> Estimate for Time 2: 100
#> Standard error:      6.93
#> Confidence interval: 86.42 to 113.58