Calculate reliable_change using various methods

This function calculates reliable_change using different methods such as Crawford, Jacobson, Speer, and others.

Usage

reliable_change(
  method = "crawford",
  t1.score,
  norm.t1.mean,
  norm.t1.sd,
  norm.t2.mean,
  norm.t2.sd,
  norm.r,
  norm.n,
  conf.level = 0.9
)

Arguments

method

The method used for reliable_change calculation. Default is "crawford".

t1.score

The score at time 1.

norm.t1.mean

Mean of the normative data at time 1.

norm.t1.sd

Standard deviation of the normative data at time 1.

norm.t2.mean

Mean of the normative data at time 2.

norm.t2.sd

Standard deviation of the normative data at time 2.

norm.r

Pearson correlation coefficient between time 1 and time 2 scores.

norm.n

Sample size of the normative data.

conf.level

Confidence level for the confidence interval calculation. Default is 0.9.

Value

A list containing the expected score at time 2, standard error of the mean, lower and upper bounds of the confidence interval.

Examples

reliable_change(method = "jacobson", 100, 103, 10, 109, 8, 0.76, 100, .95)
#> Reliable Change: Assessing if there has been a statistically significant change over time between two scores.
#> 
#> INPUT:
#> 
#> Variable    Mean   SD  Case's Score   n     r    
#> ---------  -----  ---  -------------  ----  -----
#> Time 1       103   10  100            100   0.76 
#> Time 2       109    8                            
#> 
#> PARAMS:
#> 
#> Change method:    jacobson
#> Confidence level: 0.95
#> 
#> OUTPUT:
#> Estimate for Time 2: 100
#> Standard error:      6.93
#> Confidence interval: 86.42 to 113.58